deployment/sample-agents/gcp-customer-support/requirements.txt
google-cloud-aiplatform>=1.38.0
google-cloud-logging>=3.5.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
